unix: fix uv_spawn assertion with two stdio closed
If at least two of the stdio fds are already closed before calling
`uv_spawn()`, such as stdin and stdout, then it is possible for fds 0
and 1 to be reassigned in `uv__process_init_stdio()`. If this happens,
then an assertion will occur when `uv__process_open_stream()` tries to
close one of the fds. So we should skip the stdio check.